Add files via upload
Notice Regarding jQuery Library

This project includes the jQuery library, which is used for various JavaScript functionalities.

jQuery Version
Version: 3.7.1
jQuery License
The jQuery library is licensed under the MIT License. The full license text can be found in the included jquery.js file or directly from the jQuery GitHub repository.

Redistribution Notice
By including the jQuery library in this repository:

I acknowledge the rights of the original authors of jQuery.
I have not modified the copyright or license information within the library.
I have included the library solely for ease of use and compatibility with this project.
If you wish to view, use, or modify jQuery beyond the context of this project, please refer to its official website or repository for the latest version and documentation.